Use new API v0.18 methods
For non data-streaming methods, update to use the new client method
names. For data-streaming methods, use the new adaptation functions in
`_old_component_data.py`.

Because `set_component_power_active()` can now return a the time when
the command expires, we also need to update the tests we use to run
them.

Note that the data-streaming method in the new client version is not
async (it returns a Receiver synchronously) so `await`s are also
removed.

For the `ComponentMetricFetcher` we also remove the `async` for the
wrapper `_subscribe()` methods.

The mock microgrid client was mostly rewritten to adjust it to the new
unified data streaming method.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>